Fix various DataHolder issues
While the hierarchy implementation from #27 works in some cases, it
has two problems:

- It only works for NEX version <3.5, as we aren't writing the stucture
  header. This isn't exactly a problem of the PR, as it was already
  wrong.
- It only checks for one layer of hierarchy, because there is no
  recursion applied.

Solve this by using `WriteStructure()` instead of getting the raw bytes,
since the object data is already a `StructureInterface`.

On `ExtractFromStream()` there is no need to use reflect as the object
data will be the correct structure. Using reflect also breaks with
hierarchy for unknown reasons, so just pass the dataType to
`ReadStructure()`.

Also fix a bug where the data holder is written twice when calling
`WriteDataHolder()` because of reusing the same stream on
`dataHolder.Bytes()`.