fix(datetime): Have FromTimestamp modify the underlying DateTime
This got lost in the types refactor, and was a rare case of the pointer receiver actually being needed - various apps depended on this function modifying the underlying value.

One could also argue for this being factored out into a NewDateTimeFromTimestamp function or something, but this way is source-compatible with older code.