fix(PRUDPPacketLite): Get proper SourceVirtualPortStreamID
PRUDPPacketLite has its own independent fields for the virtual ports
stream type and ID, due to them being encoded differently on the final
packet. Target those specific fields instead of the generic `sourcePort`
for consistency with the other functions and get the proper data. Fixes
client not being able to connect with PRUDPLite due to the response SYN
packet setting the destination stream ID to 0.