chore(websocket): Remove commented code
Now that the code has been tested working without using timeouts, remove
the previous code as it's not needed. Also enable back the pong sending
to be more compliant with the websocket standard. Pings are only
triggered at request by the client, so this shouldn't break any clients
dependent on the previous behavior.